# [DefaultRouter] fix unnecessary system clock reads due to races accessing router state