@wolfstudy 2021-12-27
@dferstay https://github.com/apache/pulsar-client-go/pull/684 It looks like this pr is dealing with the same thing