@dferstay 2021-12-27
@wolfstudy ,

> @dferstay #684 It looks like this pr is dealing with the same thing

Indeed, however #684 introduces a mutex on the publish path which has a negative impact on performance, especially under load from multiple concurrent go-routines.  A parallel DefaultRouter bench test was added to quantify the performance impact in #693 

For details, see my comment on #684 here: https://github.com/apache/pulsar-client-go/pull/684#discussion_r775296098